swagger integration

+++ src/utils/saveFileToCloudinary.js
@@ -1,16 +1,29 @@
 import cloudinary from 'cloudinary';
-import { env } from './env.js';
-import { CLOUDINARY } from '../constants/index.js';
+import dotenv from 'dotenv';
+import fs from 'fs/promises';
+
+dotenv.config();
 
 cloudinary.v2.config({
   secure: true,
-  cloud_name: env(CLOUDINARY.CLOUD_NAME),
-  api_key: env(CLOUDINARY.API_KEY),
-  api_secret: env(CLOUDINARY.API_SECRET),
+  cloud_name: process.env.CLOUDINARY_URL,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
+
+
 export const saveFileToCloudinary = async (file) => {
-  const response = await cloudinary.v2.uploader.upload(file.path);
-  await fs.unlink(file.path);
-  return response.secure_url;
+  try {
+    if (!file || !file.path) {
+      throw new Error('File path is missing');
+    }
+    const response = await cloudinary.v2.uploader.upload(file.path);
+    await fs.unlink(file.path);
+    return response.secure_url;
+   }catch (error) {
+    console.error('Error uploading file to Cloudinary:', error.message);
+    throw new Error('Failed to upload file to Cloudinary');
+    }
+
 };
